roadmap: move some functions out of timeline.ts
Done becaus ehte file was getting big and its good to keep it focused on drawing.
