+++ cmd/roadmap/webapp/timeline.ts
@@ -1,6 +1,6 @@
-/// <reference types="vite/client" />
-
 import p5 from "p5";
+
+import { fetchTasks, submitTasks } from "./data";
 
 let tasks: any[] = [];
 let visibleTasks: any[] = [];
@@ -22,7 +22,19 @@ let toolTipTask: any = null;
 // p5.js sketch
 const sketch = (p: p5) => {
   p.setup = async () => {
-    await fetchGanttData();
+    try {
+      tasks = await fetchTasks();
+    } catch (error) {
+      console.error("❌ Error fetching Gantt data:", error);
+    }
+    if (tasks.length === 0) {
+      console.warn("⚠️ No tasks to display.");
+      return;
+    }
+    // Initialise timeline range from the data; dateToX and the zoom init
+    // below depend on these being set before any drawing/coordinate math.
+    timelineStart = Math.min(...tasks.map((t) => new Date(t.start).getTime()));
+    timelineEnd = Math.max(...tasks.map((t) => new Date(t.end).getTime()));
 
     // Check the 'hide past tasks' checkbox
     visibleTasks = [...tasks];
@@ -130,8 +142,6 @@ const sketch = (p: p5) => {
 
     console.log("🚀 Initializing Gantt Chart...");
 
-    timelineStart = Math.min(...visibleTasks.map((t) => new Date(t.start).getTime()));
-    timelineEnd = Math.max(...visibleTasks.map((t) => new Date(t.end).getTime()));
     redraw = true;
     p.frameRate(30);
   };
@@ -720,80 +730,6 @@ function drawTooltip(p: p5, task: any, x: number, y: number) {
   p.text(`Category: ${category}`, x + padding, y + padding + 38); // ✅ Category line
 }
 
-const API_BASE_URL = import.meta.env.VITE_API_URL || "http://localhost:8080";
-
-async function fetchGanttData() {
-  try {
-    console.log("🔄 Fetching Gantt data from API...");
-
-    const response = await fetch(`${API_BASE_URL}/api/v1/timeline`);
-    console.log("API Response Received:", response);
-
-    if (!response.ok) {
-      throw new Error(`HTTP error! Status: ${response.status}`);
-    }
-
-    const rawData = await response.json();
-    console.log("📄 Raw API Data:", rawData);
-
-    // Convert API response to Task format with date parsing.
-    tasks = rawData
-      .filter((row: any) => row.Status !== "Discontinued")
-      .map((row: any) => {
-        let startDate = parseDate(row.Start || "");
-        let endDate = parseDate(row.End || "");
-        let categoryEmoji = getCategoryEmoji(row.Category || "Other");
-
-        // Validate start and end dates.
-        if (!startDate || isNaN(new Date(startDate).getTime())) {
-          console.warn(`⚠️ Invalid start date for task "${row.Title}":`, startDate);
-          startDate = new Date().toISOString().split("T")[0]; // Default to today.
-        }
-
-        if (!endDate || isNaN(new Date(endDate).getTime())) {
-          console.warn(`⚠️ Invalid end date for task "${row.Title}":`, endDate);
-          endDate = new Date(new Date(startDate).getTime() + 7 * 24 * 60 * 60 * 1000).toISOString().split("T")[0]; // Default to 7 days later
-        }
-
-        return {
-          id: row.ID,
-          category: row.Category,
-          name: `${categoryEmoji} ${row.Title}`, // Prepend emoji to task name.
-          description: row.Description || "",
-          start: startDate,
-          end: endDate,
-          priority: row.Priority || "P5",
-          owner: row.Owner || "Other",
-          milestone: row["Milestone Type"] === "Start Date" ? startDate : row["Milestone Type"] === "End Date" ? endDate : null,
-          dependencies: row.Dependencies ? row.Dependencies.split(",").map((dep) => dep.trim()) : [],
-        };
-      });
-
-    console.log("🛠️ Processed Tasks:", tasks);
-
-    // Ensure timeline range is updated.
-    if (tasks.length > 0) {
-      timelineStart = Math.min(...tasks.map((t) => new Date(t.start).getTime()));
-      timelineEnd = Math.max(...tasks.map((t) => new Date(t.end).getTime()));
-      console.log(`📆 Timeline Updated: Start - ${new Date(timelineStart).toISOString()}, End - ${new Date(timelineEnd).toISOString()}`);
-    } else {
-      console.warn("⚠️ No tasks found in API response.");
-    }
-  } catch (error) {
-    console.error("❌ Error fetching Gantt data:", error);
-  }
-}
-
-function parseDate(dateString: string): string {
-  const parts = dateString.split("/");
-  if (parts.length === 3) {
-    // Convert dd/mm/yyyy → yyyy-mm-dd
-    return `${parts[2]}-${parts[1].padStart(2, "0")}-${parts[0].padStart(2, "0")}`;
-  }
-  console.warn(`⚠️ Unexpected date format: "${dateString}"`);
-  return ""; // Return empty if the format is incorrect
-}
-
 const ownerColors: Record<string, string> = {
   "David Sutton": "rgba(255, 220, 220, 0.3)", // Light Red
   "Saxon Nelson-Milton": "rgba(220, 255, 220, 0.3)", // Light Green
@@ -826,39 +762,9 @@ function getPriorityColor(priority: string): string {
   }
 }
 
-function getCategoryEmoji(category: string): string {
-  const categoryMap: Record<string, string> = {
-    Rig: "🛰️",
-    "AusOceanTV Platform": "📺",
-    Hydrophone: "🎤",
-    Camera: "📷",
-    Broadcast: "🎥",
-    CloudBlue: "☁️",
-    Speaker: "🔊",
-    OpenFish: "🐟",
-    "Jetty Rig": "🏗️",
-    Other: "⚙️",
-  };
-
-  return categoryMap[category] || "⚙️"; // Default to "Other" emoji if no match
-}
-
 document.getElementById("submit-changes")!.addEventListener("click", async () => {
-  console.log("📤 Sending update request...", tasks);
-
   try {
-    const response = await fetch(`${API_BASE_URL}/api/v1/update`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ tasks }),
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      throw new Error(`Failed to update tasks: ${errorText}`);
-    }
-
-    console.log("Changes submitted successfully!");
+    await submitTasks(tasks);
   } catch (error) {
     console.error("❌ Error submitting changes:", error);
   }
